webroot: move duplicated code to util.js
