+++ module/webroot/index.html
@@ -17,17 +17,14 @@
             <input type="text" id="search-input" placeholder="Search for an app" autocapitalize="off">
             <button class="clear-btn" id="clear-btn">&#x2715;</button>
         </div>
-        <form id="app-form">
-            <div id="app-list">
-            <!-- System apps will be listed here -->
-            </div>
-        </form>
+        <div id="app-list">
+        <!-- System apps will be listed here -->
+        </div>
         <div class="floating-button-container">
             <div class="ripple-element" id="nuke-button">
                 <svg xmlns="http://www.w3.org/2000/svg" height="28px" viewBox="0 -960 960 960" width="28px" fill="#FFFFFF"><path d="m576-80-56-56 104-104-104-104 56-56 104 104 104-104 56 56-104 104 104 104-56 56-104-104L576-80ZM120-320v-80h280v80H120Zm0-160v-80h440v80H120Zm0-160v-80h440v80H120Z"/></svg>
             </div>
         </div>
-
         <div class="footer">
             <a class="footer-btn" id="home-btn">
                 <div class="footer-btn-icon focus-btn">
