+++ module/webroot/restore.html
@@ -17,7 +17,7 @@
             <input type="text" id="search-input" placeholder="Search for an app" autocapitalize="off">
             <button class="clear-btn" id="clear-btn">&#x2715;</button>
         </div>
-        <div id="removed-list">
+        <div id="app-list">
             <!-- Removed apps will be listed here -->
         </div>
         <div class="floating-button-container">
