+++ module/webroot/util.js
@@ -1,3 +1,5 @@
+let appList = [], nukeList = [], isShellRunning = false;
+
 export async function ksuExec(command) {
     return new Promise((resolve) => {
         let callbackName = `exec_callback_${Date.now()}`;
@@ -39,6 +41,150 @@ export function setupSearch() {
         });
         document.getElementById('clear-btn').style.display = 'none';
     });
+}
+
+// Fetch system apps
+export async function fetchAppList(file, display = false) {
+    fetch(file)
+        .then(response => response.json())
+        .then(data => {
+            data.sort((a, b) => a.app_name.localeCompare(b.app_name));
+            if (file === "app_list.json") {
+                appList = data;
+            } else {
+                nukeList = data;
+            }
+            if (display) {
+                displayAppList(data);
+                applyRippleEffect();
+            }
+        })
+        .catch(error => {
+            console.error("Failed to fetch system apps:", error);
+        });
+}
+
+// Display app list
+export async function displayAppList(data) {
+    const appListDiv = document.getElementById("app-list");
+    appListDiv.innerHTML = "";
+
+    const htmlContent = data.map((pkg) => `
+        <div class="app ripple-element" data-package-name="${pkg.package_name}" data-app-path="${pkg.app_path}">
+            <div class="app-info">
+                <img class="app-icon" src="icons/${pkg.package_name}.png" onerror="this.src='default.png'" alt="Icon">
+                <div class="app-details">
+                    <span class="app-name"><span>${pkg.app_name}</span></span>  
+                    <span class="app-package"><span>${pkg.package_name}</span></span>  
+                    <span class="app-path"><span>${pkg.app_path}</span></span>  
+                </div>
+            </div>
+            <input class="app-selector" type="checkbox">
+        </div>
+    `).join("");
+    
+    appListDiv.innerHTML = htmlContent;
+
+    // Add click handlers to all app divs
+    document.querySelectorAll('.app').forEach(appDiv => {
+        appDiv.addEventListener('click', function(e) {
+            if (e.target.type !== 'checkbox') {
+                const checkbox = this.querySelector('input[type="checkbox"]');
+                checkbox.checked = !checkbox.checked;
+            }
+        });
+
+        appDiv.querySelectorAll(".app-name span, .app-package span, .app-path span").forEach(el => {
+            const parent = el.parentElement;
+            const scrollAmount = el.scrollWidth - parent.clientWidth;
+        
+            if (scrollAmount > 0) {
+                el.classList.add("scroll");
+        
+                // Ensure the full text scrolls by adding a small buffer (+10px)
+                const adjustedScroll = scrollAmount + 10;
+                
+                // Adjust animation speed based on text length (longer text = slower scrolling)
+                const scrollTime = Math.max(3, adjustedScroll / 15); // Min 3s, longer for large text
+                
+                el.style.setProperty('--scroll-distance', `-${adjustedScroll}px`);
+                el.style.setProperty('--scroll-time', `${scrollTime}s`);
+            }
+        });
+    });
+}
+
+// Funtion to update app list
+export async function updateAppList(isNuke = false) {
+    let targetFrom, targetTo, listFrom, listTo;
+    if (isShellRunning) return;
+    if (isNuke) {
+        targetFrom = "app_list.json";
+        targetTo = "nuke_list.json";
+        listFrom = appList;
+        listTo = nukeList;
+    } else {
+        targetFrom = "nuke_list.json";
+        targetTo = "app_list.json";
+        listFrom = nukeList;
+        listTo = appList;
+    }
+
+    // Get selected apps
+    let selectedPackages = Array.from(document.querySelectorAll(".app-selector:checked"))
+        .map(checkbox => {
+            const appDiv = checkbox.closest('.app');
+            return {
+                package_name: appDiv.dataset.packageName,
+                app_path: appDiv.dataset.appPath,
+                app_name: appDiv.querySelector('.app-name').textContent
+            };
+        });
+
+    // Skip if none selected
+    if (selectedPackages.length === 0) {
+        ksu.toast("No apps selected");
+        return;
+    }
+
+    try {
+        // Add apps to targetFrom
+        const uniqueNewPackages = selectedPackages.filter(app => 
+            !listFrom.some(existingApp => existingApp.package_name === app.package_name)
+        );
+        listFrom.push(...uniqueNewPackages);
+        await ksuExec(`echo '${JSON.stringify(listFrom, null, 2)}' > /data/adb/system_app_nuker/${targetFrom}`);
+
+        // Remove apps from targetTo
+        listTo = listTo.filter(app => 
+            !selectedPackages.some(selectedApp => selectedApp.package_name === app.package_name)
+        );
+        await ksuExec(`echo '${JSON.stringify(listTo, null, 2)}' > /data/adb/system_app_nuker/${targetTo}`);
+
+        // Refresh app list
+        displayAppList(listTo);
+
+        // Update global variables
+        if (isNuke) {
+            appList = listFrom;
+            nukeList = listTo;
+        } else {
+            appList = listTo;
+            nukeList = listFrom;
+        }
+
+        // Nuke script
+        isShellRunning = true;
+        await ksuExec(`
+            PATH=/data/adb/ap/bin:/data/adb/ksu/bin:/data/adb/magisk:$PATH
+            busybox nsenter -t1 -m /data/adb/modules/system_app_nuker/nuke.sh
+        `);
+        isShellRunning = false;
+        ksu.toast("Done! Reboot your device!");
+    } catch (error) {
+        ksu.toast("Error updating removed apps list");
+        console.error("Error:", error);
+    }
 }
 
 // Function to apply ripple effect
